refactored ligature feature application to diminish duplicate code
