+++ src/features/latn/latinLigatures.js
@@ -2,39 +2,23 @@
  * Apply Latin ligature feature to a range of tokens
  */
 
-import { ContextParams } from '../../tokenizer';
-import applySubstitution from '../applySubstitution';
+import { applyFeatureToRange } from '../commonFeatureUtils';
 
 /**
- * Update context params
- * @param {any} tokens a list of tokens
- * @param {number} index current item index
+ * Apply Latin required ligatures to a context range
+ * @param {ContextRange} range a range of tokens
  */
-function getContextParams(tokens, index) {
-    const context = tokens.map(token => token.activeState.value);
-    return new ContextParams(context, index || 0);
+function latinRequiredLigature(range) {
+    applyFeatureToRange.apply(this, ['latn', 'rlig', range]);
 }
 
 /**
- * Apply Arabic required ligatures to a context range
+ * Apply Latin ligatures to a context range
  * @param {ContextRange} range a range of tokens
  */
 function latinLigature(range) {
-    const script = 'latn';
-    let tokens = this.tokenizer.getRangeTokens(range);
-    let contextParams = getContextParams(tokens);
-    contextParams.context.forEach((glyphIndex, index) => {
-        contextParams.setCurrentIndex(index);
-        let substitutions = this.query.lookupFeature({
-            tag: 'liga', script, contextParams
-        });
-        if (substitutions.length) {
-            substitutions.forEach(
-                action => applySubstitution(action, tokens, index)
-            );
-            contextParams = getContextParams(tokens);
-        }
-    });
+    applyFeatureToRange.apply(this, ['latn', 'liga', range]);
 }
 
 export default latinLigature;
+export { latinRequiredLigature, latinLigature };
